txt2img参数校验

+++ huiyu-service/service-core/src/main/java/com/huiyu/service/core/sd/dto/Txt2ImgDto.java
@@ -100,5 +100,5 @@ public class Txt2ImgDto extends Dto {
      * 种子
      */
     @JsonProperty("seed")
-    private Integer seed;
+    private Long seed;
 }
